scorer registry
