+++ llama_stack/distribution/registry/scorers/__init__.py
@@ -5,9 +5,19 @@
 # the root directory of this source tree.
 # TODO: make these import config based
 from llama_stack.apis.evals import *  # noqa: F403
+from llama_stack.providers.impls.meta_reference.evals.scorer.basic_scorers import *  # noqa: F403
 
 from ..registry import Registry
 
 
 class ScorerRegistry(Registry[BaseScorer]):
     _REGISTRY: Dict[str, BaseScorer] = {}
+
+
+SCORER_REGISTRY = {
+    "accuracy": AccuracyScorer,
+    "random": RandomScorer,
+}
+
+for k, v in SCORER_REGISTRY.items():
+    ScorerRegistry.register(k, v)
